refactor(cli): abstract file funcs to package
ALLY-844
